Refactoring (#150)
