+++ requirements.txt
@@ -1,11 +1,9 @@
 asgiref==3.4.1
 black==21.9b0
 cachetools==4.2.4
-catboost==1.0.6
 certifi==2021.10.8
 charset-normalizer==2.0.9
 click==8.0.3
-cycler==0.11.0
 Django==3.2.8
 django-debug-toolbar==3.2.2
 django-elasticsearch-dsl==7.2.1
@@ -20,7 +18,6 @@ drf-flex-fields==0.9.6
 drf-nested-routers==0.93.4
 elasticsearch==7.15.1
 elasticsearch-dsl==7.4.0
-fonttools==4.34.4
 google-api-core==2.3.0
 google-auth==2.3.3
 google-cloud-core==2.2.1
@@ -28,41 +25,25 @@ google-cloud-storage==1.43.0
 google-crc32c==1.3.0
 google-resumable-media==2.1.0
 googleapis-common-protos==1.54.0
-graphviz==0.20.1
 gunicorn==20.1.0
 idna==3.3
-joblib==1.1.0
-kiwisolver==1.4.4
-lightgbm==3.3.2
-matplotlib==3.5.2
 mypy-extensions==0.4.3
-numpy==1.23.1
-packaging==21.3
-pandas==1.4.3
 pathspec==0.9.0
-Pillow==9.2.0
+pip-autoremove==0.10.0
 platformdirs==2.4.0
-plotly==5.9.0
 protobuf==3.19.1
 psycopg2-binary==2.9.3
 pyasn1==0.4.8
 pyasn1-modules==0.2.8
 PyJWT==2.3.0
-pyparsing==3.0.9
 python-dateutil==2.8.2
 python-dotenv==0.19.1
 pytz==2021.3
 regex==2021.10.8
 requests==2.26.0
 rsa==4.8
-scikit-learn==1.1.1
-scipy==1.8.1
 six==1.16.0
-sklearn==0.0
 sqlparse==0.4.2
-tenacity==8.0.1
-threadpoolctl==3.1.0
 tomli==1.2.1
 typing-extensions==3.10.0.2
 urllib3==1.26.7
-xgboost==1.5.1
